requirements.txt
sounddevice==0.4.6
pydub==0.25.1
openai-whisper
torch
transformers
selenium
opencv-python
requests
numpy
scipy
webdriver-manager
ollama
